+++ lumos/core/graph/graph.h
@@ -32,7 +32,7 @@ Graph *create_graph();
 
 void append_layer2grpah(Graph *graph, Layer *l);
 void init_graph(Graph *g, int w, int h, int c, int truth_num, int class_num, int coretype, int subdivision, int optimizer, char *weights_path, float *input);
-void set_graph(Graph *g, float *space, int *truth, float *loss);
+void set_graph(Graph *g, float *space, float *truth, float *loss);
 void forward_graph(Graph *g, int coretype, int subdivision);
 void backward_graph(Graph *g, int coretype, int subdivision);
 void refresh_graph(Graph *g, int coretype);
