+++ lumos/core/session/session.c
@@ -25,11 +25,11 @@ void init_session(Session *sess, char *data_path, char *label_path)
 {
     if (sess->coretype == GPU){
         cudaMalloc((void**)&sess->input, sess->subdivision*sess->width*sess->height*sess->channel*sizeof(float));
-        cudaMalloc((void**)&sess->truth, sess->subdivision*sess->truth_num*sizeof(int));
+        cudaMalloc((void**)&sess->truth, sess->subdivision*sess->truth_num*sizeof(float));
         cudaMalloc((void**)&sess->loss, sizeof(float));
     } else {
         sess->input = calloc(sess->subdivision*sess->width*sess->height*sess->channel, sizeof(float));
-        sess->truth = calloc(sess->subdivision*sess->truth_num, sizeof(int));
+        sess->truth = calloc(sess->subdivision*sess->truth_num, sizeof(float));
         sess->loss = calloc(1, sizeof(float));
     }
     bind_train_data(sess, data_path);
@@ -154,25 +154,25 @@ void load_train_data_binary(Session *sess, int index)
 
 void load_train_label(Session *sess, int index)
 {
-    int *truth = calloc(sess->subdivision*sess->truth_num, sizeof(int));
+    float *truth = calloc(sess->subdivision*sess->truth_num, sizeof(float));
     for (int i = index; i < index + sess->subdivision; ++i){
-        int *truth_i = truth + (i - index) * sess->truth_num;
+        float *truth_i = truth + (i - index) * sess->truth_num;
         char *label_path = sess->train_label_paths[i];
         strip(label_path, ' ');
         void **labels = load_label_txt(label_path);
         int *lindex = (int*)labels[0];
         char *tmp = (char*)labels[1];
         for (int j = 0; j < sess->truth_num; ++j){
-            truth_i[j] = (int)atoi(tmp+lindex[j+1]);
+            truth_i[j] = atof(tmp+lindex[j+1]);
         }
         free(lindex);
         free(tmp);
         free(labels);
     }
     if (sess->coretype == GPU){
-        cudaMemcpy(sess->truth, truth, sess->truth_num*sess->subdivision*sizeof(int), cudaMemcpyHostToDevice);
+        cudaMemcpy(sess->truth, truth, sess->truth_num*sess->subdivision*sizeof(float), cudaMemcpyHostToDevice);
     } else {
-        memcpy(sess->truth, truth, sess->truth_num*sess->subdivision*sizeof(int));
+        memcpy(sess->truth, truth, sess->truth_num*sess->subdivision*sizeof(float));
     }
     free(truth);
 }
@@ -229,7 +229,7 @@ void detect_classification(Session *sess)
 {
     fprintf(stderr, "\nSession Start To Running\n");
     int num = 0;
-    int *truth = NULL;
+    float *truth = NULL;
     float *detect = NULL;
     float *loss = calloc(1, sizeof(float));
     Graph *g = sess->graph;
@@ -237,7 +237,7 @@ void detect_classification(Session *sess)
     Node *layer = g->tail;
     Layer *l = layer->l;
     if (sess->coretype == GPU){
-        truth = calloc(sess->truth_num, sizeof(int));
+        truth = calloc(sess->truth_num, sizeof(float));
         detect = calloc(sess->class_num, sizeof(float));
     }
     for (int i = 0; i < sess->train_data_num; ++i){
@@ -245,7 +245,7 @@ void detect_classification(Session *sess)
         load_train_label(sess, i);
         forward_graph(sess->graph, sess->coretype, sess->subdivision);
         if (sess->coretype == GPU){
-            cudaMemcpy(truth, l->truth, sess->truth_num*sizeof(int), cudaMemcpyDeviceToHost);
+            cudaMemcpy(truth, l->truth, sess->truth_num*sizeof(float), cudaMemcpyDeviceToHost);
             cudaMemcpy(detect, l->input, l->inputs*sizeof(float), cudaMemcpyDeviceToHost);
             cudaMemcpy(loss, sess->loss, sizeof(float), cudaMemcpyDeviceToHost);
         } else {
@@ -263,7 +263,7 @@ void detect_classification(Session *sess)
                 max = detect[j];
             }
         }
-        fprintf(stderr, "%d %d\n", truth[0], index);
+        fprintf(stderr, "%f %d\n", truth[0], index);
         if (truth[0] == index) num += 1;
         fprintf(stderr, "Loss:%.4f\n\n", loss[0]);
     }
@@ -273,7 +273,7 @@ void detect_classification(Session *sess)
 void detect_segmentation(Session *sess)
 {
     fprintf(stderr, "\nSession Start To Running\n");
-    int *truth = NULL;
+    float *truth = NULL;
     float *detect = NULL;
     int *trans = calloc(sess->truth_num, sizeof(int));
     float *loss = calloc(1, sizeof(float));
@@ -284,7 +284,7 @@ void detect_segmentation(Session *sess)
     Node *layer = g->tail;
     Layer *l = layer->l;
     if (sess->coretype == GPU){
-        truth = calloc(sess->truth_num, sizeof(int));
+        truth = calloc(sess->truth_num, sizeof(float));
         detect = calloc(sess->truth_num*sess->class_num, sizeof(float));
     }
     for (int i = 0; i < sess->train_data_num; ++i){
@@ -292,7 +292,7 @@ void detect_segmentation(Session *sess)
         load_train_label(sess, i);
         forward_graph(sess->graph, sess->coretype, sess->subdivision);
         if (sess->coretype == GPU){
-            cudaMemcpy(truth, l->truth, sess->truth_num*sizeof(int), cudaMemcpyDeviceToHost);
+            cudaMemcpy(truth, l->truth, sess->truth_num*sizeof(float), cudaMemcpyDeviceToHost);
             cudaMemcpy(detect, l->input, l->inputs*sizeof(float), cudaMemcpyDeviceToHost);
             cudaMemcpy(loss, sess->loss, sizeof(float), cudaMemcpyDeviceToHost);
         } else {
@@ -545,7 +545,7 @@ void detect_transfer(float *data, int w, int h, int c, int *trans)
     }
 }
 
-float detect_iou(int *detect, int *truth, int num)
+float detect_iou(int *detect, float *truth, int num)
 {
     int x = 0;
     for (int i = 0; i < num; ++i){
